WIP: mostly there
